Code
This code is for a project involving an ESP8266 microcontroller and ThingSpeak, an IoT (Internet of Things) platform for sending sensor data to the cloud. Specifically, the program is designed to read heartbeat data from a sensor connected to analog pin A0 and send the readings to ThingSpeak for logging and analysis.